# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: EMIRAS Stock + Vendas Inteligente API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /stock/movements:
    get:
      operationId: listStockMovements
      tags: [stock]
      summary: List stock movements
      parameters:
        - $ref: "#/components/parameters/Limit"
      responses:
        "200":
          description: Stock movements
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/StockMovement"
    post:
      operationId: createStockMovement
      tags: [stock]
      summary: Register a stock movement
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/StockMovementInput"
      responses:
        "201":
          description: Created movement
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/StockMovement"
        "400":
          $ref: "#/components/responses/BadRequest"
...
  /storage/uploads/request-url:
    post:
      tags: [Storage]
      operationId: requestUploadUrl
      summary: Request a presigned URL for file upload
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UploadUrlRequest"
      responses:
        "200":
          description: Presigned upload URL generated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UploadUrlResponse"
        "400":
          $ref: "#/components/responses/BadRequest"
        "500":
          $ref: "#/components/responses/ServerError"
...
  /storage/objects/{objectPath}:
    get:
      tags: [Storage]
      operationId: getStorageObject
      summary: Serve an uploaded object
      parameters:
        - name: objectPath
          in: path
          required: true
          schema: { type: string }
      responses:
        "200":
          description: Object content
          content:
            "*/*":
              schema: { type: string, format: binary }
        "404":
          $ref: "#/components/responses/NotFound"
components:
  parameters:
...
    Limit:
      name: limit
      in: query
      schema:
        type: integer
        minimum: 1
        maximum: 100
        default: 10
  responses:
    BadRequest:
      description: Invalid request
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
    NotFound:
      description: Resource not found
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
    ServerError:
      description: Server error
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
  schemas:
...
    Error:
      type: object
      properties:
        error:
          type: string
      required: [error]
...
    UploadUrlRequest:
      type: object
      required: [name, size, contentType]
      properties:
        name: { type: string, minLength: 1 }
        size: { type: integer, minimum: 1 }
        contentType: { type: string, minLength: 1 }
    UploadUrlResponse:
      type: object
      required: [uploadURL, objectPath]
      properties:
        uploadURL: { type: string, format: uri }
        objectPath: { type: string }
        metadata:
          $ref: "#/components/schemas/UploadUrlRequest"
    StockMovement:
      type: object
      properties:
        id: { type: integer }
        productId: { type: integer }
        productName: { type: string }
        type: { type: string, enum: [ENTRADA, SAIDA, AJUSTE, DEVOLUCAO] }
        quantity: { type: integer }
        reason: { type: string }
        previousStock: { type: integer }
        resultingStock: { type: integer }
        createdAt: { type: string, format: date-time }
      required: [id, productId, productName, type, quantity, reason, previousStock, resultingStock, createdAt]
    StockMovementInput:
      type: object
      properties:
        productId: { type: integer, minimum: 1 }
        type: { type: string, enum: [ENTRADA, SAIDA, AJUSTE, DEVOLUCAO] }
        quantity: { type: integer, minimum: 1 }
        reason: { type: string, minLength: 1 }
      required: [productId, type, quantity, reason]
...