# lib/api-zod/src/generated/types/stockMovement.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * EMIRAS Stock + Vendas Inteligente API
 * OpenAPI spec version: 0.1.0
 */
import type { StockMovementType } from './stockMovementType';

export interface StockMovement {
  id: number;
  productId: number;
  productName: string;
  type: StockMovementType;
  quantity: number;
  reason: string;
  previousStock: number;
  resultingStock: number;
  createdAt: Date;
}